fastapi_server/dev_requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
loguru
tiktoken
pydantic-settings
pytest
requests
ollama
gradio_client
